Convert Drawer to TS.
