+++ packages/design-system/src/components/Drawer/Drawer.test.tsx
@@ -10,6 +10,7 @@ const defaultProps = {
     </div>
   ),
   footerTitle: 'Footer title',
+  // eslint-disable-next-line @typescript-eslint/no-empty-function
   onCloseClick: () => {},
   heading: 'Drawer title',
 };
